+++ src/user/hooks/useUser.ts
@@ -8,7 +8,6 @@ import { useRemoteConfig } from '@/shared/hooks/useRemoteConfig';
 
 // uid로 User를 가져오는 React Query 훅 (캐시 우선)
 export function useUser(uid: string | null) {
-  const queryClient = useQueryClient();
   const noUserIdError = uid === null ? new Error('유저 ID가 존재하지 않아 유저 데이터를 불러올 수 없습니다.') : null;
 
   // 1. remote config에서 캐시 버전 받아오기
@@ -20,32 +19,25 @@ export function useUser(uid: string | null) {
   }
 
   // 3. localStorage에서 최초 데이터 읽기 (SWR: persistent cache)
-  const initialData = uid ? getCachedUserData(uid, cacheVersion) : null;
+  const safeCacheVersion = cacheVersion ?? '';
+  const initialData = uid ? getCachedUserData(uid, safeCacheVersion) : null;
 
-  // 4. storage 이벤트로 다른 탭 동기화
-  React.useEffect(() => {
-    function handleStorage(e: StorageEvent) {
-      if (e.key === `${cacheVersion}-user-${uid}`) {
-        queryClient.invalidateQueries(['user', uid]);
-      }
-    }
-    window.addEventListener('storage', handleStorage);
-    return () => window.removeEventListener('storage', handleStorage);
-  }, [uid, queryClient, cacheVersion]);
+  // 4. storage 이벤트로 다른 탭 동기화 (분리된 훅 사용, 항상 호출)
+  useUserStorageSync(uid, cacheVersion);
 
   const { data, isLoading, error } = useQuery<User | null>(
     ['user', uid],
     async () => {
-      if (uid === null) throw noUserIdError;
+      if (uid === null) throw new Error('유저 ID가 존재하지 않아 유저 데이터를 불러올 수 없습니다.');
       const user = await fetchUser(uid);
-      if (user) cacheUserData(uid, user, cacheVersion); // 최신 데이터 도착 시 localStorage 갱신
+      if (user) cacheUserData(uid, user, safeCacheVersion); // 최신 데이터 도착 시 localStorage 갱신
       return user;
     },
     {
       enabled: uid !== null,
-      initialData, // SWR: 첫 화면에 localStorage 데이터 사용
+      initialData,
       onSuccess: (user) => {
-        if (uid && user) cacheUserData(uid, user, cacheVersion); // 이중 안전
+        if (uid && user) cacheUserData(uid, user, safeCacheVersion); // 이중 안전
       },
       onError: (error) => {
         console.error('유저 데이터를 불러오던 중 에러가 발생했습니다:', error);
